feat: Enhance integrations with quantum info measures, QpiAI support, fix dtype bug
- Add 10 quantum information methods to Qiskit integration (concurrence,
  entanglement_of_formation, negativity, partial_trace, mutual_information,
  schmidt_decomposition, von Neumann entropy, state fidelity, Statevector
  from_label, DensityMatrix conversion)
- Add 8 new methods to PennyLane integration (vn_entropy, mutual_info,
  purity, fidelity, trace_distance, CHSH correlation, tomography, E91)
- Create new QpiAI Quantum SDK integration with 15 methods
- Register QpiAI in integrations __init__.py
- Fix EfficientQKDPredictor float64→float32 dtype promotion bug
- Add 25 integration tests (414 total passing)
- Clean stale files: test_results.txt, AGENTS.md, package.json/lock
- Fix docs: version in conf.py (0.2.4→0.5.0), placeholder URLs, extras
- Add enhancement plan and 3 deep framework audits (Qiskit, PennyLane, QpiAI)

+++ src/qkdpy/ml/efficient_models.py
@@ -372,10 +372,10 @@ class EfficientQKDPredictor:
         # Forward pass
         output, _, _ = self._forward(X_norm)
 
-        # Denormalize output
+        # Denormalize output (keep float32 to match internal precision)
         predictions = output.flatten() * self._output_std + self._output_mean
 
-        return predictions
+        return predictions.astype(np.float32, copy=False)
 
     def get_model_size_bytes(self) -> int:
         """Get model size in bytes."""
